Rename `pg` package to `postgres`
